Add missing declaration in Gen3

+++ src/math/Array.hpp
@@ -385,7 +385,7 @@ inline blast_fn void sincos(const Array& angles, Array& sines, Array& cosines) {
     __m256d s_tmp;
     __m256d c_tmp;
     auto vecLoopSize = (angles.size / 4) * 4;
-    for (; i < vecLoopSize; i += 4) {
+    for (; i < (int)vecLoopSize; i += 4) {
         __m256d angle_v = _mm256_loadu_pd(angles.data + i);
         s_tmp = _mm256_sincos_pd(&c_tmp, angle_v);
         _mm256_storeu_pd(sines.data+i, s_tmp);
@@ -402,7 +402,7 @@ inline blast_fn void sincos(const Array& angles, Array& sines, Array& cosines) {
     __m256 s_tmp;
     __m256 c_tmp;
     auto vecLoopSize = (angles.size / 8) * 8;
-    for (; i < vecLoopSize; i += 4) {
+    for (; i < (int)vecLoopSize; i += 4) {
         __m256 angle_v = _mm256_load_ps(angles.data+i);
         s_tmp = _mm256_sincos_ps(&c_tmp, angle_v);
         _mm256_storeu_ps(sines.data+i, s_tmp);
